+++ src/types/entry.rs
@@ -28,3 +28,26 @@ macro_rules! containing_record {
         $next.flink.cast::<u8>().sub(memoffset::offset_of!($type, $field)).cast::<$type>()
     };
 }
+
+/// Creates an iterator over doubly-linked list.
+#[macro_export]
+macro_rules! list_iter {
+    ($head:expr, $type:ty, $field:tt) => {
+        {
+            let __first = $crate::containing_record!($head, $type, $field);
+            let mut __next = __first;
+            let mut __started = false;
+            core::iter::from_fn(move || {
+                if __first == __next && __started {
+                    None
+                } else {
+                    __started = true;
+                    
+                    let val = __next.read();
+                    __next = $crate::containing_record!((*__next).$field, $type, $field);
+                    Some(val)
+                }
+            })
+        }
+    };
+}