+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
 * Added `containing_record` macro.
+* Added `list_iter` macro to iterate over doubly-linked list.
 # 0.6.0
 * Began keeping changelog ig.